update hatch code (#97)
* add new hatch mech code

* minor changes and add comments

* hardcode no delay for hatch eject

* change naming for clarity

* re-delete files from merge

* delete the imports of removed files
